Refactor solver.py into solver/ package with one file per backend
Split the monolithic solver.py (1246 lines) and qsolver.py (1191 lines)
into a solver/ package:
- solver/__init__.py: re-exports, backward-compatible public API
- solver/torch_exact.py: pure PyTorch reference solver
- solver/flash.py: Triton fused kernel solver + autograd
- solver/cutile.py: cuTile fused kernel solver + autograd
- solver/cutn.py: cuQuantum tensor network solver + helpers
- solver/qml.py: PennyLane quantum circuit solver
- solver/qiskit_solver.py: IBM Quantum via Qiskit Runtime
- solver/cudaq_solver.py: NVIDIA CUDA-Q simulation/QPU
- solver/_utils.py: shared _cast_grads_to_dtype helper

Also adds torch.dtype type hints and docstrings for c_dtype/p_dtype
in QKANLayer and QKAN constructors, documenting supported values
(torch.float32, torch.bfloat16, torch.float8_e4m3fn).

No API changes — all imports from qkan.solver continue to work.

+++ src/qkan/cutile_ops.py
@@ -191,8 +191,8 @@ def cutile_pz_forward(
     grid = (n_oi, math.ceil(batch / BLOCK_B), 1)
 
     if preacts_trainable:
-        preacts_w = preacts_w.to(c_dtype).contiguous()
-        preacts_b = preacts_b.to(c_dtype).contiguous()
+        preacts_w = preacts_w.to(io_dtype).contiguous()
+        preacts_b = preacts_b.to(io_dtype).contiguous()
 
     ct.launch(
         torch.cuda.current_stream(),
